fix(sync) Make sure we can correctly archive/delete/mark as spam
Summary:
Previously, when moving to trash or spam, we were not correctly removing
the inbox category. This produced slightly different results because the
generated task was slightly different.

Also, in ChangeMailTask we were removing the locks on the models before
the tasks had actually succeeded-- these locks allowed us to ignore
deltas on those models while the task was in progress, so removing them
before it completed caused threads to flicker when performing actions

This depends on D3829 which also properly addresses
archiving/deleting/moving to spam threads that contain sent messages

Addresses T7757 and #2706

Test Plan: unit

Reviewers: evan, halla, spang

Reviewed By: evan, halla, spang

Differential Revision: https://phab.nylas.com/D3830