[account-store] Don't `getPassword` if we don't have to
Summary:
It's slow, so only do it if we've never seen the Account before. This
fixes jank that would happen somewhat randomly. It was especially noticeable
if the user had a lot of Accounts (and therefore a lot of passwords to
load).

Test Plan: Run locally, verify that we only load new Accounts

Reviewers: spang, evan, juan

Reviewed By: evan, juan

Maniphest Tasks: T7766

Differential Revision: https://phab.nylas.com/D3848